pdf-export/requirements.txt
chardet==5.2.0
pdf2image==1.17.0
pillow==10.3.0
reportlab==4.2.0
